Added an EfficientStimuliLogger
It's probably not much more efficient but it's much simpler.
The old StimuliLogger was renamed RealTimeStimuliLogger as it logs into
files directly whereas the EfficientStimuliLogger waits for the end of
the simulation (but we can log at any point in the program before that).

EfficientStimuliLogger is great for most usages.
RealTimeStimuliLogger is better for larger simulations where holding all
the data in the RAM starts to be harder.

They both inherit from BaseStimuliLogger and this could be generalised
to loggers for data directly (AxiStreamMonitor without tlast).