Added interconnect_stream_tlast example + fixed tdest increment
tdest would increment if a ! was in the middle of a sequence because we
cut the sequences into separate Data objects during parsing assuming
they're AXI and not AXI-Stream sequences (so the address is modified).

To avoid this, we add a is_linked_to_previous flag on Data during
parsing and when the Data objects are used on an AxiStreamSource,
if they have this flag we consider they have the same address (tlast)
than the previous data in the DataList.

This is a quick fix, maybe it's better to just force the type
(AXI/AXI-Stream) when parsing.