Stimuli's Fill == -1 => generate a custom seed just for this stimuli
Before we just used the simulation current random state but that doesn't allow
to easily fix the seed to the value generated.
Now we use the current random state to generate a specific seed that is
logged, that way we can change the whole simulation seed but keep the
same seed for a Stimuli (this can allow to check if it's a specific
stimuli that causes a bug or if it's the rest of the simulation ?).