RealTimeStimuliLogger also logs axi stream without tlast stimulis
But it's highly inefficient because the Data format isn't made to be
appenned to (size in the descriptor at the beginning, to be efficient we
would need to take a padding using space characters so if the size
takes more digits we won't have to move the whole file).
The json format isn't made to be appenned to either.

Right now, we just rewrite data files entirely every time.