Removed out_axi golden stimlogs from the sizes/interconnect example.
They come from the interconnect example but since the DUT has changed
and the out_axi but doesn't exist anymore.

Maybe the tests should delete the stimlogs directory at the start of
every run if it already exists ? (each monitor deletes it's own
directory but in case a monitor is removed, its directory stays).